Created linked_view hook
