+++ lib/live_debugger_web/live/channel_dashboard_live.ex
@@ -2,32 +2,20 @@ defmodule LiveDebuggerWeb.ChannelDashboardLive do
   @moduledoc false
 
   use LiveDebuggerWeb, :live_view
+  use LiveDebuggerWeb.Hooks.LinkedView
 
   require Logger
 
-  alias LiveDebugger.Structs.LvProcess
   alias Phoenix.LiveView.JS
   alias LiveDebugger.Utils.URL
-  alias Phoenix.LiveView.AsyncResult
 
   alias LiveDebuggerWeb.Components.Error
   alias LiveDebugger.Structs.TreeNode
-  alias LiveDebugger.Services.LiveViewDiscoveryService
-  alias LiveDebugger.Utils.Parsers
-  alias LiveDebuggerWeb.Helpers.RoutesHelper
 
   alias LiveDebuggerWeb.StateLive
   alias LiveDebuggerWeb.TracesLive
   alias LiveDebuggerWeb.SidebarLive
   alias LiveDebugger.Utils.PubSub, as: PubSubUtils
-  alias LiveDebugger.Utils.Parsers
-
-  @impl true
-  def mount(%{"pid" => string_pid}, _session, socket) do
-    socket
-    |> start_async_assign_lv_process(string_pid)
-    |> ok()
-  end
 
   @impl true
   def handle_params(params, url, socket) do
@@ -95,47 +83,6 @@ defmodule LiveDebuggerWeb.ChannelDashboardLive do
     """
   end
 
-  # When fetching LvProcess fails, we try to find a successor LvProcess
-  @impl true
-  def handle_async(:fetch_lv_process, {:ok, nil}, socket) do
-    socket
-    |> handle_liveview_process_not_found()
-    |> noreply()
-  end
-
-  # When fetching LvProcess succeeds, we subscribe to its process state
-  @impl true
-  def handle_async(:fetch_lv_process, {:ok, fetched_lv_process}, socket) do
-    subscribe_process_state(fetched_lv_process.pid)
-
-    socket
-    |> assign(:lv_process, AsyncResult.ok(fetched_lv_process))
-    |> assign(:debugged_module, fetched_lv_process.module)
-    |> noreply()
-  end
-
-  # When fetching LvProcess fails, we assign the error to the socket
-  @impl true
-  def handle_async(:fetch_lv_process, {:exit, reason}, socket) do
-    Logger.error(
-      "LiveDebugger encountered unexpected error while fetching information for process: #{inspect(reason)}"
-    )
-
-    socket
-    |> assign(
-      :lv_process,
-      AsyncResult.failed(socket.assigns.lv_process, reason)
-    )
-    |> noreply()
-  end
-
-  @impl true
-  def handle_info({:process_status, :dead}, socket) do
-    socket
-    |> handle_liveview_process_not_found()
-    |> noreply()
-  end
-
   defp assign_node_id(socket, %{"node_id" => node_id}) do
     case TreeNode.id_from_string(node_id) do
       {:ok, id} ->
@@ -156,63 +103,5 @@ defmodule LiveDebuggerWeb.ChannelDashboardLive do
 
   defp assign_node_id(socket, _params) do
     assign(socket, :node_id, nil)
-  end
-
-  defp start_async_assign_lv_process(socket, string_pid) do
-    case Parsers.string_to_pid(string_pid) do
-      {:ok, pid} ->
-        socket
-        |> assign(:lv_process, AsyncResult.loading())
-        |> start_async(:fetch_lv_process, fetch_lv_process_with_retries(pid))
-
-      :error ->
-        assign(
-          socket,
-          :lv_process,
-          AsyncResult.failed(AsyncResult.loading(), :invalid_transport_pid)
-        )
-    end
-  end
-
-  defp fetch_lv_process_with_retries(pid) do
-    fn -> fetch_with_retries(fn -> LiveViewDiscoveryService.lv_process(pid) end) end
-  end
-
-  defp subscribe_process_state(pid) do
-    pid
-    |> PubSubUtils.process_status_topic()
-    |> PubSubUtils.subscribe!()
-  end
-
-  defp handle_liveview_process_not_found(socket) do
-    with %{lv_process: %{result: %LvProcess{} = lv_process}} <- socket.assigns,
-         %{pid: successor_pid} <-
-           fetch_with_retries(fn -> LiveViewDiscoveryService.successor_lv_process(lv_process) end) do
-      socket
-      |> push_navigate(to: RoutesHelper.channel_dashboard(successor_pid))
-    else
-      _ ->
-        result = %AsyncResult{
-          ok?: false,
-          loading: nil,
-          failed: :not_found,
-          result: nil
-        }
-
-        socket
-        |> assign(:lv_process, result)
-    end
-  end
-
-  defp fetch_with_retries(function) do
-    with nil <- fetch_after(function, 200),
-         nil <- fetch_after(function, 800) do
-      fetch_after(function, 1000)
-    end
-  end
-
-  defp fetch_after(function, milliseconds) do
-    Process.sleep(milliseconds)
-    function.()
   end
 end
